Small tiles are 8x8 with transparency + tweaks
Small tiles are 8x8 instead of 8x16.  Resolution of screen is too low
for big text.
Text layer pixel is transparent if set to 0.
The bottom 2K words are used for double buffering to keep the RAM
at 64KB.  The cpu can only address 63 KB of RAM.
The IO device copies the FIV register to the 2nd buffer on each frame to
keep the FIV in sync and make appear as though it is not double buffered.