feat: [timing] update timing again
Since the pc-engine/turbografx 16 had ram that ran at same speed as the
pixel clock, then we can too.
The pc-engine came out in 1987 as a mass-produced game console, I think
it is fair our console could as well.

So we re-do all the timing spec to go off of a 5.355 MHz clock
and only use 262 scan lines per frame instead of the NTSC standard of
262.5.  All the consoles of the 80's that I know of used 262 scan lines,
so let's go with that.  Otherwise, the image will appear interlaced.